# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FurniCore ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory items
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: lowStock
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Inventory list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Create inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}:
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInventoryBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventoryItem
      tags: [inventory]
      summary: Delete inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /manufacturing:
    get:
      operationId: listManufacturingTasks
      tags: [manufacturing]
      summary: List manufacturing tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: assigneeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Tasks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ManufacturingTask"
    post:
      operationId: createManufacturingTask
      tags: [manufacturing]
      summary: Create manufacturing task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateManufacturingTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManufacturingTask"
  /manufacturing/{id}:
    get:
      operationId: getManufacturingTask
      tags: [manufacturing]
      summary: Get a manufacturing task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManufacturingTask"
    patch:
      operationId: updateManufacturingTask
      tags: [manufacturing]
      summary: Update manufacturing task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateManufacturingTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManufacturingTask"
    delete:
      operationId: deleteManufacturingTask
      tags: [manufacturing]
      summary: Delete a manufacturing task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # HR - EMPLOYEES
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - name: unread
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        unit:
          type: string
        quantity:
          type: number
        reorderLevel:
          type: number
        unitCost:
          type: number
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        isLowStock:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - type
        - unit
        - quantity
        - reorderLevel
        - unitCost
        - isLowStock
        - createdAt
        - updatedAt
    CreateInventoryBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        unit:
          type: string
        quantity:
          type: number
        reorderLevel:
          type: number
        unitCost:
          type: number
        supplierId:
          type: integer
      required:
        - name
        - type
        - unit
        - quantity
        - reorderLevel
        - unitCost
    UpdateInventoryBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        unit:
          type: string
        quantity:
          type: number
        reorderLevel:
          type: number
        unitCost:
          type: number
        supplierId:
          type: integer

    # SUPPLIERS
...
    ManufacturingTask:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: ["integer", "null"]
        productName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        assigneeId:
          type: ["integer", "null"]
        assigneeName:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: string
        estimatedHours:
          type: ["number", "null"]
        actualHours:
          type: ["number", "null"]
        progress:
          type: integer
        dueDate:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - status
        - priority
        - progress
        - createdAt
        - updatedAt
    CreateManufacturingTaskBody:
      type: object
      properties:
        productId:
          type: integer
        title:
          type: string
        description:
          type: string
        assigneeId:
          type: integer
        priority:
          type: string
        estimatedHours:
          type: number
        dueDate:
          type: string
      required:
        - title
        - priority
    UpdateManufacturingTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        assigneeId:
          type: integer
        status:
          type: string
        priority:
          type: string
        estimatedHours:
          type: number
        actualHours:
          type: number
        progress:
          type: integer
        dueDate:
          type: string

    # HR - EMPLOYEES
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
        isRead:
          type: boolean
        link:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - userId
        - title
        - message
        - type
        - isRead
        - createdAt

    # ACTIVITY LOGS
...